- Issue Fixed

+++ lib/src/ui/knowledge/knowledge_list_screen.dart
@@ -198,7 +198,7 @@ class _KnowledgeListScreenState extends State<KnowledgeListScreen> {
                                     context,
                                     KnowledgeDetailScreen.routeName,
                                     arguments: item,
-                                  ) as int;
+                                  );
                                 },
                               ),
                             ),
